Fix imports for tests
In stable rustc the `use pom::Parser` and `use pom::parser::*` conflict on the `Parser` types. `pom::Parser`
is just a helper alias which is the same as `parser::Parser` with a `'static` lifetime, so for tests just
use that explicitly.

It would be better to have a longer-term fix, probably by renaming one of the `Parser` types.